first pass. All envs in colosseum_v2 dir are running
