+++ mani_skill/envs/tasks/tabletop/colosseum_v2_versions/colosseum_v2_env_utils.py
@@ -7,7 +7,7 @@ REALSENSE_DEPTH_FOV_HORIZONTAL_RAD = 87.0 * np.pi / 180
 
 SHADER = "default"
 
-def get_camera_configs(xy_offset, z_offset, target: tuple[float, float, float], camera_width, camera_height):
+def get_camera_configs(xy_offset: float, z_offset: float, target: tuple[float, float, float], camera_width: int, camera_height: int):
     pose_center = sapien_utils.look_at(eye=[xy_offset, 0,  z_offset], target=target)
     pose_left = sapien_utils.look_at(eye=[0.0, -xy_offset, z_offset], target=target)
     pose_right = sapien_utils.look_at(eye=[0.0, xy_offset, z_offset], target=target)
